Remove code that handles Mac line endings
This block intended to convert old Mac line endings ('\r', CR, x0D) to
Unix line endings ('\n', LF, x0A). The problem is that it inadvertently
replaced these bytes in the middle of lines, which can easily occur with
non latin character sets in different Unicode encodings.

This fixes
https://sourceforge.net/p/kdiff3/bugs/198/

This patch is a modified version of the one at
https://sourceforge.net/u/bluemonkmn/kdiff3/ci/7406acd54e0558b9941123154dcfb161f78fda01/